rename new components
