+++ src/fast_math.cpp
@@ -1,0 +1,1 @@
+#include "fast_math.hpp"
