+++ src/lib_dispatch.cpp
@@ -1,4 +1,4 @@
-#include "blas_wrapped.hpp"
+#include "lib_dispatch.hpp"
 #include <iostream>
 #include <type_traits>
